Extract CompactSortedFilesTestBase
